+++ src/main/typescript/config.ts
@@ -4,6 +4,8 @@ import WvActorSheet from "./applications/actor/wvActorSheet.js";
 import EffectSheet from "./applications/item/effectSheet.js";
 import WeaponSheet from "./applications/item/weaponSheet.js";
 import { CONSTANTS, TYPES } from "./constants.js";
+import { JSON_SCHEMA as BASE_ITEM_JSON_SCHEMA } from "./data/item/baseItem.js";
+import { WEAPON_SOURCE_JSON_SCHEMA } from "./data/item/weapon/source.js";
 import { getGame } from "./foundryHelpers.js";
 import WvCombat from "./foundryOverrides/wvCombat.js";
 import WvRuler from "./foundryOverrides/wvRuler.js";
@@ -14,13 +16,22 @@ import {
   flagCriticalFailure,
   flagCriticalSuccesses
 } from "./rolls/criticalsModifiers.js";
+import { JSON_SCHEMA as RULE_ELEMENT_JSON_SCHEMA } from "./ruleEngine/ruleElementSource.js";
 import { initializedSettingName } from "./settings.js";
 
 /** The Foundry configuration function for the init hook */
 export function configureFoundryOnInit(): void {
+  const ajv = new Ajv({ allErrors: true });
   getGame().wv = {
-    ajv: new Ajv({ allErrors: true }),
-    macros
+    ajv,
+    macros,
+    validators: {
+      item: {
+        effect: ajv.compile(BASE_ITEM_JSON_SCHEMA),
+        weapon: ajv.compile(WEAPON_SOURCE_JSON_SCHEMA)
+      },
+      ruleElement: ajv.compile(RULE_ELEMENT_JSON_SCHEMA)
+    }
   };
 
   // Register our own Document classes.